EthCrack - Master
Hi all, I've been working on a tool for brute-forcing ETH private keys. The main purpose of this tool is to contribute to the effort of solving the Ethereum puzzle transactions.

It is open-source under the MIT licence and requires no external dependencies other than the CUDA toolkit. It builds on Windows using Visual Studio 2015, and Linux using Make (you might have to edit the Makefile and point it towards your CUDA toolkit directory). It can search for compressed/uncompressed keys or both.

The performance is good, but can likely be improved. On my hardware (GeForce GT 640) it gets 9.4 million keys per second compressed, 7.3 million uncompressed.

Note:
-Currently it is CUDA only.
-It can only search one target key at a time


Features I would like to add if there is enough interest for the project:

-Support for searching multiple target keys at one time Done
-OpenCL/AMD device support
-CPU with AVX/AVX2/SHA support
-Checkpoints/Stop and resume
-Vanity address generation

Source and Win32/Win64 binaries available here: